+++ CHANGELOG.md
@@ -4,6 +4,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased] - 2023-07-29
+### Added
+- Compatibility with Detail Armor Bar.
+
+### Changed
+- Grouped gilded armor with the other armor sets in the creative tab.
+
 ## [1.4.1] - 2023-01-24
 ### Changed
 - Gilded enderite armor can only be repaired with the ingot from its own mod.